+++ drizzlepac/haputils/align_utils.py
@@ -4,8 +4,8 @@ import copy
 import sys
 import traceback
 import warnings
+from packaging.version import Version
 
-from distutils.version import LooseVersion
 from collections import OrderedDict
 
 import numpy as np
@@ -68,7 +68,7 @@ class AlignmentTable:
         * **apply_fit** : Updates all input image WCSs with the result of the selected 'best' fit
 
     """
-    def __init__(self, input_list, clobber=False, dqname='DQ',
+    def __init__(self, input_list, clobber=False, dqname='DQ', process_type='',
                  log_level=logutil.logging.NOTSET, **alignment_pars):
         """
         Parameters
@@ -82,6 +82,10 @@ class AlignmentTable:
         dqname : str, optional
             Allows the user to customize the name of the extension (`extname`) containing the
             data quality flags to be applied to the data during source identification.
+
+        process_type : str, optional
+            Specifies what type of data processing is being done on the input data.
+            Values include: '' (default for pipeline processing), 'SVM', 'MVM'.
 
         log_level : int, optional
             Set the logging level for this processing
@@ -141,7 +145,7 @@ class AlignmentTable:
         # Apply filter to input observations to insure that they meet minimum criteria for being able to be aligned
         log.info(
             "{} AlignmentTable: Filter STEP {}".format("-" * 20, "-" * 63))
-        self.filtered_table = analyze.analyze_data(input_list, type="SVM")
+        self.filtered_table = analyze.analyze_data(input_list, type=process_type)
         log.debug("Input sorted as: \n{}".format(self.filtered_table))
 
         if self.filtered_table['doProcess'].sum() == 0:
@@ -664,7 +668,7 @@ class HAPImage:
 
         # astropy's code returned the opposite bitmask from what was originally
         # defined by stsci.tools own bitmask code.
-        if LooseVersion(stsci.tools.__version__) >= '4.0.0':
+        if Version(stsci.tools.__version__) >= Version('4.0.0'):
             dqmask = ~dqmask
 
         return dqmask
@@ -1218,9 +1222,14 @@ def update_image_wcs_info(tweakwcs_output, headerlet_filenames=None, fit_label=N
                     wcs_name = '{}-FIT_{}_{}'.format(wname, fit_label, item.meta['fit_info']['catalog'])
 
             # establish correct mapping to the science extensions
-            sci_ext_dict = {}
-            for sci_ext_ctr in range(1, num_sci_ext + 1):
-                sci_ext_dict["{}".format(sci_ext_ctr)] = fileutil.findExtname(hdulist, 'sci', extver=sci_ext_ctr)
+            try:
+                sci_ext_dict = {}
+                for sci_ext_ctr in range(1, num_sci_ext + 1):
+                    sci_ext_dict["{}".format(sci_ext_ctr)] = fileutil.findExtname(hdulist, 'sci', extver=sci_ext_ctr)
+            except Exception:
+                exc_type, exc_value, exc_tb = sys.exc_info()
+                traceback.print_exception(exc_type, exc_value, exc_tb, file=sys.stdout)
+                logging.exception("message")
 
         # update header with new WCS info
         sci_extn = sci_ext_dict["{}".format(item.meta['chip'])]