+++ drizzlepac/haputils/generate_custom_svm_param_file.py
@@ -51,7 +51,6 @@ SPLUNK_MSG_FORMAT = '%(asctime)s %(levelname)s src=%(name)s- %(message)s'
 log = logutil.create_logger(__name__, level=logutil.logging.NOTSET, stream=sys.stdout,
                             format=SPLUNK_MSG_FORMAT, datefmt=MSG_DATEFMT)
 __version__ = 0.1
-__version_date__ = '04-Dec-2020'
 # ----------------------------------------------------------------------------------------------------------------------