+++ drizzlepac/haputils/product.py
@@ -137,7 +137,7 @@ class HAPProduct:
 
         return meta_wcs
 
-    def align_to_gaia(self, catalog_list=[], output=True,
+    def align_to_gaia(self, catalog_list=[], output=True, process_type='SVM',
                       fit_label='SVM', align_table=None, fitgeom=''):
         """Extract the flt/flc filenames from the exposure product list, as
            well as the corresponding headerlet filenames to use legacy alignment
@@ -183,6 +183,7 @@ class HAPProduct:
                 # If necessary, generate the alignment table only once
                 if align_table is None:
                     align_table = align_utils.AlignmentTable(exposure_filenames,
+                                                             process_type=process_type,
                                                              log_level=self.log_level,
                                                              **alignment_pars)
                     align_table.find_alignment_sources(output=output, crclean=crclean)
@@ -287,13 +288,25 @@ class HAPProduct:
                                 alignment_pars['determine_fit_quality']['do_consistency_check'] = False
 
                                 # Evaluate the quality of the fit
+                                # Need to create this log file specifically for the PyTest testing environment
+                                # when a single Python session is running and the tests are looking for dataset-specific
+                                # log files.
+                                log_file = "temp_align.log"
                                 is_good_fit, _, _, _, _, _ = align.determine_fit_quality_mvm_interface(align_table.imglist,
                                                                                                        align_table.filtered_table,
                                                                                                        more_catalogs,
                                                                                                        num_cat,
                                                                                                        alignment_pars,
                                                                                                        print_fit_parameters=True,
-                                                                                                       loglevel=self.log_level)
+                                                                                                       loglevel=self.log_level,
+                                                                                                       runfile=log_file)
+                                
+                                # Clean up the temporary log file as the contents are captured
+                                try:
+                                    os.remove(log_file)
+                                except OSError as error:
+                                    log.warning("Unable to remove file {}.".format(log_file))
+                                    log.warning("Output trailer files may contain duplicate information.")
 
                                 # Ensure the original parameters stay intact for the iterations
                                 # as the perform_fit() modifies the fitgeom
@@ -315,8 +328,9 @@ class HAPProduct:
                             except Exception:
                                 log.info("Problem with fit done for catalog '{}' with method '{}' using fit geometry '{}'.".
                                          format(catalog_item, method_name, mosaic_fitgeom))
-                                traceback.print_exc()
-
+                                exc_type, exc_value, exc_tb = sys.exc_info()
+                                traceback.print_exception(exc_type, exc_value, exc_tb, file=sys.stdout)
+                                logging.exception("message")
 
                             # Try again with a different fit geometry algorithm
                             mosaic_fitgeom_index -= 1